+++ wxc_sdk/as_api.py
@@ -62,8 +62,8 @@ __all__ = ['AsAccessCodesApi', 'AsAdminAuditEventsApi', 'AsAgentCallerIdApi', 'A
            'AsForwardingApi', 'AsGroupsApi', 'AsGuestManagementApi', 'AsHotelingApi', 'AsHuntGroupApi',
            'AsIncomingPermissionsApi', 'AsInternalDialingApi', 'AsJobsApi', 'AsLicensesApi',
            'AsLocationAccessCodesApi', 'AsLocationInterceptApi', 'AsLocationMoHApi', 'AsLocationNumbersApi',
-           'AsLocationVoicemailSettingsApi', 'AsLocationsApi', 'AsManageNumbersJobsApi', 'AsMeetingChatsApi',
-           'AsMeetingClosedCaptionsApi', 'AsMeetingInviteesApi', 'AsMeetingParticipantsApi',
+           'AsLocationVoicemailSettingsApi', 'AsLocationsApi', 'AsMSTeamsSettingApi', 'AsManageNumbersJobsApi',
+           'AsMeetingChatsApi', 'AsMeetingClosedCaptionsApi', 'AsMeetingInviteesApi', 'AsMeetingParticipantsApi',
            'AsMeetingPreferencesApi', 'AsMeetingQandAApi', 'AsMeetingQualitiesApi', 'AsMeetingTranscriptsApi',
            'AsMeetingsApi', 'AsMembershipApi', 'AsMessagesApi', 'AsMonitoringApi', 'AsMusicOnHoldApi', 'AsNumbersApi',
            'AsOrganisationVoicemailSettingsAPI', 'AsOrganizationApi', 'AsOutgoingPermissionsApi', 'AsPagingApi',
@@ -7542,6 +7542,67 @@ class AsIncomingPermissionsApi(AsPersonSettingsApiChild):
         await self.put(ep, params=params, data=settings.model_dump_json())
 
 
+class AsMSTeamsSettingApi(AsApiChild, base='telephony/config/people'):
+    async def read(self, person_id: str,
+             org_id: str = None) -> MSTeamsSettings:
+        """
+        Retrieve a Person's MS Teams Settings
+
+        At a person level, MS Teams settings allow access to retrieving the `HIDE WEBEX APP` and `PRESENCE SYNC`
+        settings.
+
+        To retrieve a person's MS Teams settings requires a full or read-only administrator auth token with a scope
+        of `spark-admin:telephony_config_read`.
+
+        :param person_id: Unique identifier for the person.
+        :type person_id: str
+        :param org_id: ID of the organization in which the person resides. Only admin users of another organization
+            (such as partners) may use this parameter since the default is the same organization as the token used to
+            access the API.
+        :type org_id: str
+        :rtype: :class:`MSTeamsSettings`
+        """
+        params = {}
+        if org_id is not None:
+            params['orgId'] = org_id
+        url = self.ep(f'{person_id}/settings/msTeams')
+        data = await super().get(url, params=params)
+        r = MSTeamsSettings.model_validate(data)
+        return r
+
+    async def configure(self, person_id: str,
+                  setting_name: str, value: bool,
+                  org_id: str = None):
+        """
+        Configure a Person's MS Teams Setting
+
+        MS Teams settings can be configured at the person level.
+
+        Requires a full administrator auth token with a scope of `spark-admin:telephony_config_write`.
+
+        :param person_id: Unique identifier for the person.
+        :type person_id: str
+        :param setting_name: The enum value to be set to `HIDE_WEBEX_APP`.
+        :type setting_name: str
+        :param value: The boolean value to update the `HIDE_WEBEX_APP` setting, either `true` or `false`. Set to `null`
+            to delete the `HIDE_WEBEX_APP` setting.
+        :type value: bool
+        :param org_id: ID of the organization in which the person resides. Only admin users of another organization
+            (such as partners) may use this parameter since the default is the same organization as the token used to
+            access the API.
+        :type org_id: str
+        :rtype: None
+        """
+        params = {}
+        if org_id is not None:
+            params['orgId'] = org_id
+        body = dict()
+        body['settingName'] = setting_name
+        body['value'] = value
+        url = self.ep(f'{person_id}/settings/msTeams')
+        await super().put(url, params=params, json=body)
+
+
 class AsMonitoringApi(AsPersonSettingsApiChild):
     """
     API for person's call monitoring settings
@@ -9108,6 +9169,8 @@ class AsPersonSettingsApi(AsApiChild, base='people'):
     hoteling: AsHotelingApi
     #: Person's Monitoring Settings
     monitoring: AsMonitoringApi
+    #; MS Teams settings
+    ms_teams: AsMSTeamsSettingApi
     #: music on hold settings
     music_on_hold: AsMusicOnHoldApi
     #: Phone Numbers for a Person
@@ -9147,6 +9210,7 @@ class AsPersonSettingsApi(AsApiChild, base='people'):
         self.forwarding = AsPersonForwardingApi(session=session)
         self.hoteling = AsHotelingApi(session=session)
         self.monitoring = AsMonitoringApi(session=session)
+        self.ms_teams = AsMSTeamsSettingApi(session=session)
         self.music_on_hold = AsMusicOnHoldApi(session=session)
         self.numbers = AsNumbersApi(session=session)
         self.permissions_in = AsIncomingPermissionsApi(session=session)